Moved Thread methods, and added deletion handling - Added `ThreadMarkerRepository` for managing thread creation, updates, and deletion. - Introduced support for "Deleted" marker status. - Updated controllers and other repositories to use `ThreadMarkerRepository` methods. - Cleaned up redundant thread management code in `MapMarkerRepository` and `MapMarkerSuggestionRepository`.
