+++ MarkerStatus.php
@@ -7,16 +7,22 @@ enum MarkerStatus: string
     case ACTIVE = 'Active';
     case INACTIVE = 'Inactive';
     case DISABLED = 'Disabled';
+    case DELETED = 'Deleted';
 
     /**
      * Determine the status of a marker based on its properties
      *
      * @param bool $active Whether the marker is active
      * @param bool $createThread Whether the marker has create_thread enabled
+     * @param bool $isDeleted Whether the marker is marked for deletion
      * @return self The appropriate status enum value
      */
-    public static function fromMarker(bool $active, bool $createThread): self
+    public static function fromMarker(bool $active, bool $createThread, bool $isDeleted = false): self
     {
+        if ($isDeleted) {
+            return self::DELETED;
+        }
+
         if (!$active) {
             return self::DISABLED;
         } elseif (!$createThread) {
