+++ Repository/MapMarkerRepository.php
@@ -4,13 +4,10 @@ namespace Sylphian\Map\Repository;
 
 use Exception;
 use Sylphian\Map\Entity\MapMarker;
-use Sylphian\Map\MarkerStatus;
 use XF;
-use XF\Entity\Thread;
 use XF\Mvc\Entity\AbstractCollection;
 use XF\Mvc\Entity\Repository;
 use XF\PrintableException;
-use XF\Service\Thread\CreatorService;
 
 /**
  * Repository for map marker operations
@@ -145,6 +142,13 @@ class MapMarkerRepository extends Repository
     {
         try {
             $marker = $this->getMapMarkerOrFail($id);
+
+            if ($marker->thread_id) {
+                /** @var ThreadMarkerRepository $threadMarkerRepo */
+                $threadMarkerRepo = $this->repository('Sylphian\Map:ThreadMarker');
+                $threadMarkerRepo->markThreadAsDeleted($marker);
+            }
+
             $marker->delete();
             return true;
         } catch (Exception $e) {
@@ -312,113 +316,5 @@ class MapMarkerRepository extends Repository
             'markerTypes' => array_values($markerTypes),
             'allMarkers' => $allMarkers
         ];
-    }
-
-    /**
-     * Creates a thread for a map marker
-     *
-     * @param MapMarker $marker The marker to create a thread for
-     * @param string|null $customTitle Optional custom title for the thread (defaults to marker title)
-     * @return bool Whether the thread was successfully created
-     */
-    public function createThreadForMarker(MapMarker $marker, ?string $customTitle = null): bool
-    {
-        if (!XF::options()->enableThreadCreation) {
-            return false;
-        }
-
-        $threadCreationLocation = XF::options()->threadCreationLocation;
-        if (!$threadCreationLocation) {
-            return false;
-        }
-
-        $forum = XF::em()->find('XF:Forum', $threadCreationLocation);
-        if (!$forum) {
-            XF::logError('Thread creation failed: Forum not found with ID ' . $threadCreationLocation);
-            return false;
-        }
-
-        try {
-            /** @var CreatorService $creator */
-            $creator = XF::service('XF:Thread\CreatorService', $forum);
-
-            //TODO: To be changed at some point.
-            // Placeholder until I figure out how to create prefixes on addon instillation.
-            $status = MarkerStatus::fromMarker($marker->active, $marker->create_thread);
-
-            $baseTitle = $customTitle ?: $marker->title;
-            $formattedTitle = "[{$status->value}] {$baseTitle}";
-
-            $creator->setContent(
-                $formattedTitle,
-                $this->getThreadMessageFromMarker($marker)
-            );
-
-            $errors = [];
-            if ($creator->validate($errors)) {
-                $thread = $creator->save();
-
-                $marker->thread_id = $thread->thread_id;
-                $marker->save();
-
-                return true;
-            } else {
-                XF::logError('Thread creation validation failed: ' . implode(', ', $errors));
-                return false;
-            }
-        } catch (Exception $e) {
-            XF::logException($e, false, 'Error creating thread for map marker: ');
-            return false;
-        }
-    }
-
-    /**
-     * Updates the title of a thread associated with a marker
-     *
-     * @param MapMarker $marker The marker with an associated thread
-     * @return bool Whether the update was successful
-     * @throws PrintableException
-     */
-    public function updateThreadTitle(MapMarker $marker): bool
-    {
-        if (!$marker->thread_id) {
-            return false;
-        }
-
-        /** @var Thread $thread */
-        $thread = XF::em()->find('XF:Thread', $marker->thread_id);
-        if (!$thread) {
-            return false;
-        }
-
-        $status = MarkerStatus::fromMarker($marker->active, $marker->create_thread);
-
-        $pattern = MarkerStatus::getRegexPattern();
-        $currentTitle = $thread->title;
-
-        if (preg_match('/^\[(' . $pattern . ')\] (.+)$/i', $currentTitle, $matches)) {
-            $baseTitle = $matches[2];
-        } else {
-            $baseTitle = $currentTitle;
-        }
-
-        $thread->title = "[{$status->value}] {$baseTitle}";
-        $thread->save();
-
-        return true;
-    }
-
-    /**
-     * Generates thread message content from a map marker
-     *
-     * @param MapMarker $marker
-     * @return string
-     */
-    protected function getThreadMessageFromMarker(MapMarker $marker): string
-    {
-        return "This thread is associated with a map marker:\n\n" .
-            "Title: {$marker->title}\n" .
-            "Description: {$marker->content}\n" .
-            "Location: {$marker->lat}, {$marker->lng}";
     }
 }
